requirements.txt
appdirs==1.4.3
boto==2.46.1
Django==1.11
django-bootstrap3==8.2.2
django-ckeditor==5.2.2
django-crispy-forms==1.6.1
django-embed-video==1.1.2
django-extensions==1.7.5
django-password-reset==1.0
django-reversion==2.0.8
django-storages==1.5.2
djangorestframework==3.5.3
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.10.3
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pyparsing==2.2.0
PyPDF2==1.26.0
pytz==2017.2
reportlab==3.4.0
requests==2.13.0
reversion==0.2
six==1.10.0
toml==0.9.2
webencodings==0.5.1
xhtml2pdf==0.2b1
